refactor enhancements
